post in draft order and select order from product detail

+++ app/src/main/java/com/example/mcommerceapp/pojo/draft_orders/LineItems.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class LineItems (
 
   @SerializedName("id"                   ) var id                 : Int?                = null,
-  @SerializedName("variant_id"           ) var variantId          : Int?                = null,
+  @SerializedName("variant_id"           ) var variantId          : Long?                = null,
   @SerializedName("product_id"           ) var productId          : Int?                = null,
   @SerializedName("title"                ) var title              : String?             = null,
   @SerializedName("variant_title"        ) var variantTitle       : String?             = null,
